// app/src/store/mutation-types.ts
export enum MutationTypes {
  EXAMPLE = 'EXAMPLE',

  CURRENT_SPACE_SET = 'CURRENT_SPACE_SET',
  CURRENT_TRAIL_SET = 'CURRENT_TRAIL_SET',

  CURRENT_TREE_SET = 'CURRENT_TREE_SET',
  CURRENT_TREE_CONFIG_SET = 'CURRENT_TREE_CONFIG_SET',

  S3_CONFIG_SET = 'S3_CONFIG_SET',

  LOADING_STATE_SET = 'LOADING_STATE_SET',

  TROVE_STATE_SET = 'TROVE_STATE_SET',
}